[PROJECT] - Refacto project with RemoteLoader helper to avoid duplication code
